Follow up to storybook upgrade - make filenames consistent
Also gets rid of some new typescript syntax which prettier v1 can't handle
