+++ src/BloomBrowserUI/bookEdit/StyleEditor/fontSelectComponent.stories.tsx
@@ -1,5 +1,4 @@
-/** @jsx jsx **/
-import { jsx, css } from "@emotion/react";
+import { css } from "@emotion/react";
 import * as React from "react";
 import FontSelectComponent from "./fontSelectComponent";
 import FontInformationPane from "../../react_components/fontInformationPane";
