+++ src/BloomBrowserUI/package.json
@@ -190,7 +190,7 @@
         "less-loader": "^11.1.0",
         "less-watch-compiler": "^1.13.0",
         "lessc-glob": "^1.0.9",
-        "lint-staged": "^13.0.3",
+        "lint-staged": "^15.4.3",
         "lorem-ipsum": "^2.0.2",
         "markdown-it-attrs": "^2.3.3",
         "markdown-it-container": "^2.0.0",
@@ -234,6 +234,6 @@
     "packageManager": "yarn@1.22.19",
     "volta": {
         "node": "22.11.0",
-        "yarn": "1.22.19"
+        "yarn": "1.22.22"
     }
 }
