+++ src/BloomBrowserUI/problemDialog/NotifyDialog.stories.tsx
@@ -1,30 +1,10 @@
-import { ProblemKind } from "./ProblemDialog";
 import { NotifyDialog } from "./NotifyDialog";
-import { ReportDialog } from "./ReportDialog";
 
 const message =
     "Fake error with a line break<br> and <b>bold</b> and <a href='https://google.com'>link</a>...";
 
 export default {
-    title: "Problem Report"
-};
-
-export const FatalError = () => <ReportDialog kind={ProblemKind.Fatal} />;
-
-FatalError.story = {
-    name: "FatalError"
-};
-
-export const NonFatalError = () => <ReportDialog kind={ProblemKind.NonFatal} />;
-
-NonFatalError.story = {
-    name: "NonFatalError"
-};
-
-export const UserProblem = () => <ReportDialog kind={ProblemKind.User} />;
-
-UserProblem.story = {
-    name: "UserProblem"
+    title: "Problem Report/NotifyDialog"
 };
 
 export const NotifyUserNonReportable = () => (
