+++ src/BloomBrowserUI/publish/stories/accessibility.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import type { Meta, StoryObj } from "@storybook/react";
+import { Meta, StoryObj } from "@storybook/react";
 import { AccessibilityCheckScreen } from "../accessibilityCheck/accessibilityCheckScreen";
 import "../storiesApiMocks";
