+++ src/BloomBrowserUI/publish/stories/base-template.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import type { Meta, StoryObj } from "@storybook/react";
+import { Meta, StoryObj } from "@storybook/react";
 import PublishScreenTemplate from "../commonPublish/PublishScreenTemplate";
 import PublishScreenBanner from "../commonPublish/PublishScreenBanner";
 import { Button, Typography } from "@mui/material";
