+++ src/BloomBrowserUI/publish/stories/bloom-reader.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import type { Meta, StoryObj } from "@storybook/react";
+import { Meta, StoryObj } from "@storybook/react";
 import { ReaderPublishScreen } from "../ReaderPublish/ReaderPublishScreen";
 import "../storiesApiMocks";
