+++ src/BloomBrowserUI/publish/stories/video.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import type { Meta, StoryObj } from "@storybook/react";
+import { Meta, StoryObj } from "@storybook/react";
 import { PublishAudioVideo } from "../video/PublishAudioVideo";
 import "../storiesApiMocks";
