+++ src/BloomBrowserUI/react_components/color-picking/Colors.stories.tsx
@@ -1,6 +1,5 @@
 // Don't add /** @jsxFrag React.Fragment */ or these stories won't show up in StoryBook! (at least in Aug 2022)
-/** @jsx jsx **/
-import { jsx, css } from "@emotion/react";
+import { css } from "@emotion/react";
 
 import * as React from "react";
 import { useState } from "react";
