+++ src/BloomBrowserUI/react_components/stories/book-information-cards.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { BookInfoCard } from "../bookInfoCard";
 
-import type { Meta, StoryObj } from "@storybook/react";
+import { Meta, StoryObj } from "@storybook/react";
 
 const meta: Meta = {
     title: "BookInformationCards"
@@ -35,9 +35,7 @@ export const PreviouslyUploadedStory: Story = {
 
 export const NewUploadStory: Story = {
     name: "NewUpload",
-    render: () => (
-        <BookInfoCard title="02. Foo Bar" languages={languages2} />
-    )
+    render: () => <BookInfoCard title="02. Foo Bar" languages={languages2} />
 };
 
 export const SeveralLanguagesStory: Story = {
