+++ src/BloomBrowserUI/react_components/stories/dialogs.stories.tsx
@@ -13,7 +13,7 @@ import {
     normalDialogEnvironmentForStorybook
 } from "../BloomDialog/BloomDialogPlumbing";
 
-import type { Meta, StoryObj } from "@storybook/react";
+import { Meta, StoryObj } from "@storybook/react";
 
 const meta: Meta = {
     title: "Misc/Dialogs"
